login check added to endpoints - need to fix all endpoint tests

+++ src/main/java/alignpro/Controller/ProjectOverview/ProjectController.java
@@ -2,6 +2,7 @@ package alignpro.Controller.ProjectOverview;
 
 import alignpro.Model.Projects.Project;
 import alignpro.Service.ProjectService;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -18,20 +19,23 @@ public class ProjectController {
 
 
     @GetMapping("/CreateProject")
-    public String createProject(Model model){
+    public String createProject(Model model, HttpSession session) {
+        if (isUserLoggedIn(session)) return "redirect:/login";
         Project obj = new Project();
         model.addAttribute("obj", obj);
         return "create-Project";
     }
 
     @PostMapping("/saveProject")
-    public String saveProject(@ModelAttribute Project newProject){
+    public String saveProject(@ModelAttribute Project newProject, HttpSession session) {
+        if (isUserLoggedIn(session)) return "redirect:/login";
         projectService.saveProject(newProject);
         return "redirect:/";
     }
 
     @GetMapping("/edit-project/{projectId}")
-    public String editProject(@PathVariable("projectId") int projectId, Model model) {
+    public String editProject(@PathVariable("projectId") int projectId, Model model, HttpSession session) {
+        if (isUserLoggedIn(session)) return "redirect:/login";
         Project project = projectService.getProject(projectId);
         if (project != null) {
             model.addAttribute("project", project);
@@ -46,14 +50,17 @@ public class ProjectController {
                                 @RequestParam("projectName") String projectName,
                                 @RequestParam("projectDescription") String projectDescription,
                                 @RequestParam("startDate") String startDate,
-                                @RequestParam("deadLine") String deadLine) {
+                                @RequestParam("deadLine") String deadLine,
+                                HttpSession session) {
+        if (isUserLoggedIn(session)) return "redirect:/login";
         Project project = new Project(projectId, projectName, startDate, deadLine, projectDescription);
         projectService.editProject(project, project.getProjectID());
         return "redirect:/";
     }
 
     @PostMapping("/delete-project/{projectID}")
-    public String deleteProject(@PathVariable("projectID") int projectID){
+    public String deleteProject(@PathVariable("projectID") int projectID, HttpSession session) {
+        if (isUserLoggedIn(session)) return "redirect:/login";
         Project projectToDelete = projectService.getProject(projectID);
         if(projectToDelete != null){
             projectService.deleteProject(projectID);
@@ -61,6 +68,10 @@ public class ProjectController {
         } else {
             return "redirect:/";
         }
+    }
+
+    public boolean isUserLoggedIn(HttpSession session){
+        return session.getAttribute("pmUserID") == null;
     }
